setActionEvent sur le bouton ajouterLivraison, reordonnencement des methodes dans les etats
