+++ project/src/main/java/org/apache/project/controleur/Controleur.java
@@ -67,6 +67,10 @@ public class Controleur {
 		etatCourant.calculerTournee(this, planDeVille, demandeDeLivraison, tournee, fenetrePrincipale);
 	}
 	
+	public void ajouterLivraison() {
+		etatCourant.ajouterLivraison(this, planDeVille, demandeDeLivraison, tournee, fenetrePrincipale);
+	}
+	
 	public void intersectionClicked(Intersection intersection) {
 		etatCourant.intersectionClicked(this,  planDeVille,  demandeDeLivraison,  tournee,  fenetrePrincipale,  intersection);
 	}
