+++ project/src/main/java/org/apache/project/controleur/EtatAjoutLivraison1.java
@@ -1,6 +1,7 @@
 package org.apache.project.controleur;
 
 import org.apache.project.modele.DemandeDeLivraison;
+import org.apache.project.modele.Intersection;
 import org.apache.project.modele.PlanDeVille;
 import org.apache.project.modele.Tournee;
 import org.apache.project.vue.FenetrePrincipale;
@@ -8,7 +9,8 @@ import org.apache.project.vue.FenetrePrincipale;
 public class EtatAjoutLivraison1 extends EtatDefaut{
 	
 	@Override
-	public void ajouterLivraison(Controleur controleur, PlanDeVille planDeVille, DemandeDeLivraison demandeDeLivraison, Tournee tournee, FenetrePrincipale fenetrePrincipale) {
-		controleur.setEtatCourant(controleur.etatAjoutLivraison2);
+	public void intersectionClicked (Controleur controleur, PlanDeVille planDeVille, DemandeDeLivraison demandeDeLivraison, Tournee tournee, FenetrePrincipale fenetrePrincipale, Intersection intersection) {
+		controleur.setEtatCourant(controleur.etatAjoutLivraison3);
+	
 	}
 }
