+++ project/src/main/java/org/apache/project/controleur/EtatAjoutLivraison2.java
@@ -2,6 +2,7 @@ package org.apache.project.controleur;
 
 import org.apache.project.modele.DemandeDeLivraison;
 import org.apache.project.modele.Intersection;
+import org.apache.project.modele.Livraison;
 import org.apache.project.modele.PlanDeVille;
 import org.apache.project.modele.Tournee;
 import org.apache.project.vue.FenetrePrincipale;
@@ -9,10 +10,14 @@ import org.apache.project.vue.FenetrePrincipale;
 
 public class EtatAjoutLivraison2 extends EtatDefaut{
 	
-	@Override
-	public void intersectionClicked (Controleur controleur, PlanDeVille planDeVille, DemandeDeLivraison demandeDeLivraison, Tournee tournee, FenetrePrincipale fenetrePrincipale, Intersection intersection) {
-		controleur.setEtatCourant(controleur.etatAjoutLivraison3);
+private Intersection intersectionLivraison;
 	
+	public void livraisonClicked(Controleur controleur, PlanDeVille planDeVille, DemandeDeLivraison demandeDeLivraison, Tournee tournee, FenetrePrincipale fenetrePrincipale, Livraison livraisonPrecedente) {
+		
+	}
+
+	protected void actionEntree(Intersection intersection) {
+		this.intersectionLivraison = intersection;
 	}
 
 }
