+++ project/src/main/java/org/apache/project/controleur/EtatInit.java
@@ -26,7 +26,7 @@ public class EtatInit extends EtatDefaut {
 			e.printStackTrace();
 		} catch (ExceptionXML e) {
 			e.printStackTrace();
-			fenetrePrincipale.afficherPopup(e.getMessage());
+			fenetrePrincipale.afficherPopupError(e.getMessage());
 		}
 	}
 }
